Enum 파일 분리
